+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **MINOR BREAKING CHANGES**
   - Entity instances should never be compared by reference equality; always do
     `entity1.equals(entity2)`.
+  - It is no longer valid to additional managers or systems to initialized worlds.
 
 - MundaneWireException thrown by World#inject if class lacks @Wire.
   Previously, it silently failed to inject.